+++ frontend/src/types/Exercise.ts
@@ -1,0 +1,10 @@
+import Set from './Set'
+
+interface Exercise {
+  exerciseName: string
+  sets: Set[]
+  rest: string
+  weight?: number
+}
+
+export default Exercise
