+++ frontend/src/types/LocalProgram.ts
@@ -1,0 +1,8 @@
+import Exercise from './Exercise'
+
+interface LocalProgram {
+  programName: string
+  exercises: Exercise[]
+}
+
+export default LocalProgram
