+++ frontend/src/types/Set.ts
@@ -1,0 +1,6 @@
+interface Set {
+  setId: number
+  reps: number
+}
+
+export default Set
